Move ProjectFieldKeys into jira_types.py
